WL#15524 Patch #8 TLS options in ndb_mgm
Support the standard --ndb-tls-search-path in ndb_mgm.

Also support the --ndb-mgm-tls option, but in ndb_mgm, this enum
option can take a third value, "deferred", that is not supported
by other programs.

Add a --test-tls option, which makes a single, verbose attempt to
connect using TLS, then exits with status 0 on success or 1 on error.

Add a new interactive command, "START TLS". This can perform the
TLS negotiation that was deferred with --ndb-mgm-tls=deferred.

Change-Id: I711245b723183ffe97f87dd0d7d15d2349730c77