WL#15524 Patch #15 Visbility
Add functions and struct to MGM API:
  int ndb_mgm_has_tls()
  struct ndb_mgm_cert_table
  int ndb_mgm_list_certs()
  void ndb_mgm_cert_table_free()

Add a "list certs" request and reply to the management protocol.

In the management server's reply to the "get session" and
"list sessions" commands, add a row indicating whether a session
has TLS.

In the ndb_mgm client, add the "TLS INFO" command.
Test this command in a new MTR test, tls_info.

Change-Id: If91c730a743b55f2b6ed805caec45a2c19de23b5