Bug#35629957 routertest_component_gr_state QuorumTest fails randomly.
The test is launching the Router with various quorum scenarios. Regardless
whether the Router has access to quorum or not the test always uses
metadata-cache refresh cycles to synchronize. This sometimes fails
when the Router is expect to start accepting ports.

This patch changes the test to wait for the Router notify "READY" message
whenever the accepting ports are expected to be open instead of using refresh
cycles for synchronization.

Change-Id: I4524656da571a3be6cf4b4ddaa087033c8fe43de