WL#15902 Tablesample Support for HeatWave
What is Tablesample:
======================
The SQL standard provides two sampling methods; SYSTEM and BERNOULLI.

SYSTEM is basically page-level sampling, where the storage engine
returns pages instead of rows. For HeatWave SYSTEM sampling implies
chunk level granularity i.e. if a chunk is decided to be
returned, ALL rows in that chunk are returned.

BERNOULLI is row-level sampling, where the storage engine has to look
at each row individually in order to determine whether or not to
return that row.

SYNTAX:
================
SELECT select_expr FROM table_references
TABLESAMPLE { SYSTEM | BERNOULLI } ( sample_percentage )

TABLESAMPLE is only applicable for the base table and when the query
is executed in HeatWave, otherwise suitable error should be thrown.

Execution in PRIMARY and SECONDARY engine
=========================================
TABLESAMPLE can be executed only in Heatwave. In case the query cannot
be executed in HeatWave or the secondary engine is switched OFF, then
it cannot be executed in MySQL either since MySQL does not support
sampling. The error message will be "ERROR HY000: Can be executed in
Secondary engine only due to the presence of 'TABLESAMPLE'"

High level Design:
=================
SYSTEM sampling is also referred to as chunk-level sampling. In this
approach, a random number is generated for each chunk. If the
generated random number is less than or equal to the designated
sampling percentage, the entire chunk is selected for subsequent
processing. This process occurs while iterating through the chunks
of the base tables. To ensure this functionality during query
execution, the System tablesample information is integrated into
the qkrn node itself.

Bernoulli sampling is also recognized as row-level sampling. In this
method, a random number is generated for each row. If the generated
random number is less than or equal to the specified sampling
percentage, the row is chosen for subsequent processing. In practice,
Bernoulli sampling is translated into the condition
"rand() < sampling_percentage" and incorporated into the table scan
filter condition.

Bug#35859409 - WL#15902: Bernoulli sampling works with 0.0 to 1.0
                         but not with 0 to 100
Bug#35870474 - WL#15902: Mysqld crash - Assertion
`path->materialize().table_path->type == Acce
Bug#35859424 - WL#15902: System sampling always returns full table
Bug#35899074 - WL#15902: Mysqld crash at MoveCompositeIteratorsFromTablePath

Change-Id: I6475838aa2ce66d7fc10ef169219aa7023bac2c5