Bug#35634714: Derived condition pushdown return wrong results
When pushing a condition down to a derived table, we clone the
condition. While cloning, if the underlying field is a view
reference (field from a merged derived table), we strip off
the view reference and clone the expression that it references.
For the case in bugpage, this underlying expression happens
to be a constant expression from a table that is on the inner
side of an outer join. A constant expression in such a
case cannot be treated as a basic constant because of the
"NULLs" that need to be generated and this is handled carefully
in Item_view_ref::used_tables(). So when we strip off the view
reference, this information is lost and therefore we see wrong
results.
So the fix is to avoid condition pushdown for such a case.

Change-Id: I0d09d73649b7a9b5c22b27156b40dd56027e255d