Bug#33301070 wrong error-msg if config-file has wrong permissions
The config-file given to mysqlrouter can't be read, mysqlrouter reports:

  The configuration file '...' does not exist.

This is caused by mysql_harness::Path::exist() not being able to
distingiush between "file-not-found" and "other-error".

Additionally, in windows _stat() returns file-not-found if the file
exists, but can't be read. Using GetFileAttribute() properly reports if
the file exists.

Change
------

- fixed file-type check on windows to return a proper error in case a
  file isn't readable by using GetFileAttribute() instead of stat()
- added test on Unix

Change-Id: Ibfb72e5d0a65d0924532f50c8ab422e747cecddb