WL#15841 Option to allow traffic when the sole reachable partition has no quorum
This WL allows the user to configure the Router so that it keeps the accepting
ports open even if it only has an access to the nodes of the GR that have no
quorum.

The new supported router option from the metadata is
"unreachable_quorum_allowed_traffic".
It can be set to one of the following:
- "none" - current (default) behavior where the Router shuts down all the
accepting ports if there is no quorum
- "read" - Router keeps the RO and RW-split accepting ports open
- "all" - Router keeps all the accepting ports open

Change-Id: Ib7218edaac82f6d981a734e2c312901d4568971e