Bug#35345903: MySQL 8.0.33 crashes at copy_columns
Background:
  WL#11605 enabled converting character sets with ALGORITHM=INPLACE
when the source character set is a subset of destination character
set. When InnoDB is converting columns, it creates the required
heap only when old table is in compact format and new table is
non-compact format. It expects that converting from redundant to
redundant does not require any heap as column length will remain
equal.

Issue:
  After the worklog was implemented, the function to copy columns
in case of converting row formats was re-used to convert character
sets. However, in case the heap was not allocated, it was expected
that there is no change in length in case of redundant. As shown
in the test case, this is not the case for converting character set.

Fix:
  Generalize the heap creation to cases when the new table is in
redundant format irrespective of the old table. Error out in case
the heap is not allocated and new table is redundant.

Change-Id: Id304c394c6953f8b04a2ed4aab3c4d04a7ae1b75