Bug#35652019 mock keeps running after bootstrap_failover()
When the component tests call bootstrap_failover, the mock-server that
is started by bootstrap_failover keeps running until the test ends,
instead of "until bootstrap is done".

That blocks the TCP ports which may be used by a 2nd run of the
mock-server on the same port in the same test.

Change
------

- shutdown the mock-servers used by bootstrap_failover()

Change-Id: I8962f311afb5d32fea0703fbd102aea73274235a