Bug#35254025: Mysql instance cannot be initialized correctly   because autocommit is turned off.
Turning autocommit to off (from the default on) will mess up
with the --initialize code.
There's no point in honoring autocommit in that mode.
Turned off auto-commit if --initialize is also specified.
Added a warning on ignoring autocommit in that mode to the user.
Added a test.
Turned off TLS for --initialize to save some time from generating
certs.

Change-Id: I97f49a6d0d020360c63f840c6cae69d0f8d25d0b