Bug#35150382: Failure in sp_prepare_func_item() function
This is a regression from a performance enhancement in bug#32742537.
When Item_field::save_in_field() is called, some information is cached
so that if the value of the same Item_field is copied to the same Field,
if the previous copy was determined to be a simple memcpy(), so is the
next copy.

However, the cache validity check was insufficient. In some cases,
such as when pushing an argument onto a call stack before calling a
stored function, the address of the Field for two subsequent copy
operations might be the same, even though the Field objects are actually
different. This is what happens in the original test case for
the statements STORED_B(_parameter1) and STORED_C(_parameter1).

The solution to the problem is to apply more checks before reusing
cached values.

- The value to copy must be the same size. This fixes the original
  bug report, but notice that there may still be problems returning
  wrong values when the fields have different type information.

- The from and to fields must have the same type. This fixes a problem
  where the first field copied to is a BIGINT and the second field is
  a DOUBLE. These fields have the same size, but a conversion is needed.

- For a DECIMAL field, the scale must be the same. A DECIMAL(10, 2)
  has the same size as DECIMAL(10, 0), but a conversion is needed.

Character field copies are never cached, so there's no reason to
worry about those.

Change-Id: I09b68e122fa128cc19b49d23d919d3f6017f2a65