Bug#35763094 Refactor ConfigInfo::ParamInfo constructors
Add specific constructor for CI_SECTION type parameters which overrides
the type for _default member and do not use _min and _max members.

Add specific constructor for CI_ENUM type parameters which overrides the
type for _min member and do not use _max member.

This makes the construction more type safe.

And also reduces the need for explicit casts in other parts of code.

Change-Id: I8212be107cc2c02d50d110080f1053b404974ce4