Bug#35653673 Clean up Transporter interfaces after introduction of MultiTransporter
Patch is a pure refactoring of the transporter interfaces
such that all its implementations use only the TrpId (transporter id) to
identify the Transporter to be used , as well as addressing the send buffers to
append the sent messageges into.

It cleans up obfuscation added by:

------------
commit bac0d05d61e68237c35eed0399e36eaea297471f
Author: Mikael Ronström <mikael.ronstrom@oracle.com>
Date:   Tue Oct 8 19:18:32 2019 +0200

    WL#12837: Multiple transporters between nodes in a node group
------------
That work added an extra TrpId argument to these interfaces, as it was
required to address one of possibly multiple transporters to a specific NodeId.
However, this change was not carried through for the API-transporter inteface
implementations where only a single transporter could be used.
Thus, some level of obfuscation were added.

Some sloopy usage of just specifying a 'Uint32' where either a NodeId or
a TrpId should be used is also fixed.

Note that:

- The method performSendNode() became obsolete -> removed
- The methods getSendBufferLevel() were never used.
  This patch only comments out these methods as we expect a variant
  of these methods to be used in some upcomming patches.

Change-Id: I2b33f74856a59c5c025352963f0f5d328f62ab57