Bug#35102220: Assertion failure in                  Item::walk_helper_thunk at ../sql/item.h
Post-push fix: The test case for Bug#34716246 failed with
hypergraph and valgrind post this bugfix. This is because
the underlying expression for the Item_ref involved in
condition pushdown had multiple fields for which the
"depended_from" and "context" had to be copied. However,
post the bugfix for Bug#35102220, after we copy the
"depended_from" and "context" for the first field item,
we cleared the information. There by making the next field
in the expression not an outer reference. This lead to
problems while planning in hypergraph - specifically
trying to fetch the used tables for the table.

So the fix now is to make a copy of the "depended_from"
and "context" set by the Item_ref object. This way, we
can always use the saved depended_from and context info
for the next field under the same Item_ref object.

Change-Id: I7f0528b06227cfd173616cf75e2b2505daa9755b