Bug#35328028: prepared stmt of setting sp_var from subquery fails
This problem is a regression from the solution of bug#33754993.
The problem was that a CREATE TABLE statement that failed during
preparation was not properly cleaned up after execution, so that the
next execution of the same statement might fail critically.
The solution to this problem was to explicitly invalidate a statement
that failed during preparation.

However, the SET statement is not represented by a proper Sql_cmd class
and thus does not have the state management that other DML statements
have. Thus, a failure during opening of tables for an already prepared
statement would seem to be a failure during preparation. This would
further cause the function bind_tables() to be applied on non-prepared
items, which is a no-op, and a subsequent privilege check on the same
item could lead to accessing a null pointer.

There is however a more elegant solution to the CREATE TABLE problem, and
that is to set up a Prepare_error_tracker object when preparing
CREATE TABLE. This causes the statement object to be properly marked
as invalid, and a subsequent execution will clean out and reprepare the
statement. This makes it possible to revert the code for explicit
invalidation mentioned above.

Thank you to Lu Hao for providing this contribution.

Change-Id: I52edc93506e8862c8f1b605e182916323cab6ffc