Bug#34704094 - Mysql str_to_date can correct convert for '2022-11-30'
Issue:
STR_TO_DATE function doesn't validate the date as whole, it takes
date like '2022-11-31' as a valid datetime.

Analysis:
When we check date time value for validity, we need to check if the
day mentioned in the query is not greater than total number of days
for the given month. This check was missing because of which invalid
date like '2022-02-31' was getting considered as correct.

Fix:
Fixed the issue by providing above mentioned check.

Change-Id: Ibeecbbbc28eea137ce0cf01a87301905fba6bd53