WL#15786 Automatically updated histograms
This patch adds automatically updated histograms to the server.

Individual histograms can be set to be either automatically or manually
updated by specifying a new trailing option [{AUTO|MANUAL} UPDATE] on
the UPDATE HISTOGRAM command. The default setting is MANUAL UPDATE.

Automatic updates of histograms take place as follows:

When ANALYZE TABLE <table> is called automatically updated histograms on
<table> will be updated (re-created with the same settings from a fresh
sample of data from the table) and made available to the optimizer.

For InnoDB tables, when automatic recalculation of persistent statistics
takes place, any automatically updated histograms on the table are also
updated and made available to the optimizer.

Change-Id: I4755c80a896962a6094a6bf5140477927d3a2e31