Bug#35671595: Prepared statement with CTE assertion error
When a parameter is cloned E.g as part of re-parsing a common table
expression or re-parsing a condition to pushdown to derived table,
it's always added as a clone to the original parameter in the query.
To add as a clone in case of condition pushdown, we use the position
of the original parameter in the query. However, when a parameter that
needs to be cloned is already a clone created by reparsing a CTE,
position of this parameter does not reflect the position of it's master.
As a result, while searching for adding the newly created param as a
clone to the master, because of the mismatch in the position, master
param was never found resulting in failing an assert.
Fix is to search the list of the clones for each of the param in the
original query.
This fixes the problem with prepare. However, when this query needs
to be re-prepared during execution because of the type not being set,
query would fail. After the clone is created, it's param_state is not
synced up with it's master which makes Item_param::fix_fields() set
the data_type to MYSQL_TYPE_INVALID. So we make an additional fix to
sync up clones before the param gets resolved.

Change-Id: Icb9ee6260dac26112d336ecf12cf3d140ec2cb1b