Bug#35664004 Transporter level errors in NdbApi stdout/stderr
In transporter socket authentication, suppress the "negotiation failed"
error message that can occur as a normal part of node startup. The
remaining conditions that might be logged are all unexpected conditions.

The bug raises some issues that are not addressed here, including the
general question of where these error messages are directed to, and the
fact that the message does not provide any context about the identity
of the peer node.

Change-Id: Ie99348110e80b7c1a663f61da3c0c3e862b07765