Bug#35472091 : TABLE_SHARE::is_secondary_engine get_eligible_secondary_engine_from
Bug#35553940 : Segmentation fault in TABLE_SHARE::is_primary_engine in sql/table.h

When an error is thrown in Sql_cmd_dml::execute function, we know there
is already a preparation or execution failure at this point. There is
no need to ask for secondary_engine_fail_reason at this stage, it should
be sufficiently complete. We should rather ask for this kind of
information just after an offload failure has occurred, which we already
do in Bug#35493481 (forced secondary engine execution goes to primary
engine and picks wrong error).

Change-Id: I7c568d4c2024fac3e45b29b5a7ef4412ecbcd11a