Bug#35689804: Depending on the data inserted, the results of the               PREPARE and the general query run are different.
When a condition which is being pushed down to a derived table is cloned
and if any of the expressions in the condition have parameters in the
underlying derived table, for some cases, type propagation was not
happening (even after the call to fix_fields()). It is expected that a
call to propagate_type() is made for such cases. This was missed.
So we fix this by making a call to propagate_type() now.

Change-Id: I1191aaa3f095dd54451667631678260c0169e764