Bug#35700332 Handling of ENOMEM in ::doSend() should be removed
Handling of ENOMEM errors returned from writev() and recv() in resp.
TCP_Transporter::doSend() and TCP_Transporter::doReceive() was introduced
by the patch:

.......................

    commit 367d9278a182676d8ba6a4d768ae9e3eae05c440
    Author: Mikael Ronström <mikael.ronstrom@oracle.com>
    Date:   Fri Jun 28 14:44:26 2019 +0200

        BUG#29887068
    ....
    Handling ENOMEM with resends of smaller size when send attempted.
......................

The ENOMEM handling attempted a resend with smaller size by direcly
reducing the iov_len of the first iovec[] buffer.

At best this seemed to have solved a corner case where we would
run out of memory anyway real soon.

The testcase which reportedly was solved was a manually run attempt
to test many-nodes functionality on a small set of machines, perhaps
explaining the memory pressure issues. Lots of others overload issues
were also observed in similar tests, which could be claimed to use
a configuration we really do not recomend nor support

In lack of evidence that the ENOMEM handling had any value, and as
we would like to keep the code 'clean', this patch revert the ENOMEM
handling introduced by the patch refered to above.

Change-Id: Ib5a269f0b4a41a6d683d27f49b575616d82d3822