Bug#34929930: Sub-statement of a stored procedure support in secondary engine.
This is a followup patch to fix a regression from the first patch.

Due to a problem with one of the tests in rapid.sp that was discovered
during development of the first patch, some code that was removed from
sp_lex_instr::validate_lex_and_execute_core() in the fix for
bug#35328028 was added back. This fixed the actual problem, but
naturally reintroduced the problem from the other bug.

However, the actual problem was due to inconsistent state management in
open_secondary_engine_tables(). When a stored procedure with a SELECT
statement was run in forced secondary mode, and that statement had
a reference to a stored function, the statement was correctly flagged as
being disabled for execution in secondary engine. But the state was
still set to run in SECONDARY. The execution code got the message and
reported an error that the statement could not run in secondary storage
engine due to containing references to stored functions,

However, for the second execution, open_secondary_engine_tables() would
see that statement was disabled from running in secondary engine, and
then set it to run in the primary engine, and the execution was
successful, even though it should not be.

The solution to this is to improve on the state management in
open_secondary_engine_tables() and not force the state change to PRIMARY
when SECONDARY is forced.

We also extend documentation to notify the special case where
forced secondary execution is requested but secondary tables cannot be
used, but nevertheless is the statement left with opened primary tables.

Change-Id: Ide5cf67381cdd7529f9f8c35edfc926123b58e03