Bug#35702079 Need to keep transporter sendbuffers updated during onging disconnect
If TCP_transported::doSend() encounted an error requiring a
disconnect to be initiated, it returned immediately without
updating the send buffers with the data it had already sent.
It also returned 'true', which indicate that doSend() need to
be retried.

As transporter disconnection is an asynchronous process we can
not rule out that another doSend() will be retried before the
disconnection has reached a state where more sends will be prevented.
(Not sure if that will ever succeed though)

In the meantime it is a good idea to still keep the send buffers
updated.

Patch breaks out of the send loop instead of doing an immediate return.
Thus the code maintaining the send buffers with 'sum_sent' will be
executed.

Change-Id: Id81134c9db670041613cd02641a7963d43766a3c