Bug#35705590 Router crashes when target is clusterset and only has access to no-quorum part.
New tests for "WL#15841 Option to allow traffic when the sole reachable
partition has no quorum" revealed a legacy issue where the Router crashes if
the target Cluster of the Router is part of a ClusterSet and the Router only has
an access to the sub-group of this cluster with no GR quorum.

The problem is the access to std::optional<> member without checking it is set,
which happens in case of Router having only access to no-quorum group.

This patch fixes the issue checking if the optional value is set before
accessing it.

This scenario is tested by WL#15841 test ClusterSetAccessToPartitionWithNoQuorum.

Change-Id: I90b3400ed5e08e03699dbb6d252755cfc93ad75f