WL#15696 Firewall: Periodically reload data from tables
Firewall stores permanent data in tables. Data is read during firewall plugin
startup and stored in internal memory cache. Modification of the data outsidee
of the plugin is not reflected in the internal until the plugin is reloaded
(uninstall/install plugin or server restart).

This feature introduces a method for periodic reload of data from tables.

Change-Id: Iae7e9a34bb9f7ca7d94a13696a720f7cac88e21b