Bug#35710378: mysql server 8.1.0 failed at Query_expression::num_visible_fields
This problem is similar to a problem that was fixed in WL#9571,
and it appeared because of changes to evaluation sequence applied in
fix for bug#35184353.

The problem may occur when we have a subquery predicate that is
converted to an anti-join, and one of the expressions involved in the
subquery predicate is a scalar subquery, and the condition is optimized
away because it is always false. The scalar subquery remains in the
sj_outer_exprs list and is later accessed when attempting to optimize
semi-join nests for materialization.

Similar problems for a semi-join is handled correctly by clearing the
SJ expression lists when the semi-join condition is false. However,
the condition for analysis for anti-join is not set up the same.

The fix for the problem consists of several steps:

First, the main problem is fixed by analyzing the proper anti-join
condition for a constant value.

Next, we realize that we can eliminate expressions from the semi-join
expression lists both when the condition has been determined to be
always false and always true (earlier, only the "always false" case
was handled), since in both cases the semi-join expressions need not be
evaluated.

Next, the logic in Query_block::build_sj_cond() has been slightly
simplified by eliminating all constant predicates, and adding a
constant predicate unconditionally when all predicates are constant.

Finally, this might cause a fairly large number of test result
differences. The number of differences was reduced by returning a
boolean "simple_const" from build_sj_cond(), which is true when the
function installs a single Item_func_true or Item_func_false as the
semi-join condition, and do not eliminate the semi-join expressions
when this is true.

Change-Id: I486d48b396f5152dc19a9be82ff2ded8901c1fb0