Bug#34306497 - Low limit heuristic used unnecessarily with descending scans
Issue:
The low limit heuristic is not working well for order by desc as it
chooses the wrong index.

Analysis:
Because of the low limit we relook at the access method that is chosen
which in this case is the index. If the chosen range scan can provide
the order we skip the heuristic. At this point since const optimization
is not done the optimizer thinks the current range scan cannot provide
order so it relooks at other options. Since the primary key can provide
the order it chooses that one.

Fix:
Since const optimisation is not done when we check if an index can
provide order, the key part of the index and the order by field did not
match. The fix is to check if the index can provide the order by
skipping constant key parts preceding to the specified order by
expression. Later we check if the best range scan can provide order by
reversing it.

Change-Id: I1f967670c195f49fe39284e17e2d3c70030ad466