Bug#35439787 Assertion `false && "Inconsistent row counts for different AccessPath objects."'
There was an error in AddPredicate() so that it would always estimate
the selectivity of a condition in the context of the root
CompanionSet. This lead to inconsistent row estimates, as the
selectivity was in other cases estimated using the correct
CompanionSet. This commit fixes that by ensuring that AddPredicate()
picks the CompanionSet for the tables in that condition.

Change-Id: I40156663f32a407490b066a566abe72b356cfda9