Bug#35737357 Router creates the metadata user with 'mysql_native_password'
When the Router is bootstrapped against the Cluster without
existing account given via --account parameter it auto-creates
a metadata user. If the --force-password-verification parameter
is not given, the auto-created user is using mysql_native_password
auth plugin. This plugin is deprecated in 8.1 and about to be
removed in the future.

This ptach changes the behavior so that the Router is using the default
authentication plugin in the Cluster as if --force-password-validation was
given.

Change-Id: I41a235daeb4bba67fd7640b0865ce93f68083346