Bug#35737521 connect fails with preferred/as_client and old clients
Authenticating via non-ssl connections and using 3rd party clients
fails with:

- Couldn't read RSA public key from server

when using caching-sha2-password's fast-auth support and:

- client_ssl_mode=PREFERRED or PASSTHROUGH
- server_ssl_mode=AS_CLIENT

The problem does not appear with:

- libmysqlclient based clients
- if SSL is used
- if another client_ssl_mode/server_ssl_mode combination is used.

Root Cause
==========

The error is generated by the client as it interprets the protocol
differently than libmysqlclient based clients.

With those client_ssl_mode/server_ssl_mode combinations the router
wronly assumes that it should ask the client to switch the authentication
method although forwarding the server's response would be the correct
message flow.

It only affects PREFERRED/PASSTHROUGH + AS_CLIENT as they have different
starting protocol flow which does NOT ask the client for the plaintext
password over secure transports.

But the AuthForwarder assumes that the password has already been asked
for and gets out of sync.

Change
======

- track if the client's plaintext password was requested idependently
  from the state of "in-handshake".

Change-Id: I2de5e0d0b3fc2a8eaa193096cb32a681500af78a