Bug#35703114 Assertion `m_psi_batch_mode == PSI_BATCH_MODE_NONE' failed.
The fix for bug#34940000 "Hash join execution may be inefficient if
probe input is empty" caused a regression. An assert failure occurs
under the following circumstances:
- We execute a hash join where we start reading from the probe input
  e.g. a left join.
- There is an error when reading that first row, e.g. because there
  is a filter on the probe input where we get a string conversion error.

In this situation, we stop executing the join, but do not turn off PFS
batch mode for the probe iterator. (PFS batch mode is an optimization
for queries involving the performance schema.) This triggers the
assert. This commit fixes this bug by ensuring that we stop PFS batch
mode in this situation.

Change-Id: I49d03d1896f3cff7aa01058e194be2afb89a0f81