WL#15605 MLE Milestone 2
Component Service Function to getting and changing the MLE Stored Program from the sp_head.
Currently the sp_head handles the lifetime of the MLE Stored Program, creating it when needed, and destroying it
when the sp_head gets destroyed.
But we need to be able to manually destroy the MLE Stored Program, and notify the sp_head to detach it, when:
 * aborting the SP execution
 * Resetting the MLE Session when incompatible changes are made (e.g. timezone changes)
 * Manual user reset of the Session
 * exceeding the number of active MLE SPs
 * etc.
Using the `set()` notion providing nullptr to detach.
Making sure that the sp_head provided, is part of the current_thd. Documenting it.

Change-Id: I45ed2f4a56058d8bed24ae9885c14deb92b07344